rewrote grammar
I decided to include entanglement with measurement and corrections, as opposed to being split out like the preparations, because the paper seemed to list examples with entanglement sandwiched between other commands; entanglement could then be moved to the front via equivalence relations. If we do decide to split it out, it wouldn't be too difficult at all to do so.