fixed utils
Notably, added a get_outcome parameter to new_angle and calc_signal. get_outcome is a function and should be passed as an argument when calling these functions. get_outcome takes a qubit as input and returns the outcome, 0 or 1. Since getting the outcome of a qubit will depend on the implementation of the system in question, this function must be written in each implementation.